Skip empty map when enable partal update in RedisSink

+++ java/feathub-connectors/flink-connectors/flink-connector-redis/src/main/java/com/alibaba/feathub/flink/connectors/redis/sink/RedisSinkFunction.java
@@ -99,7 +99,9 @@ public class RedisSinkFunction extends RichSinkFunction<RowData> {
                                 data.getMap(valueFieldIndex), (KeyValueDataType) dataType);
 
                 if (newMap == null || newMap.isEmpty()) {
-                    client.del(key);
+                    if (!enableHashPartialUpdate) {
+                        client.del(key);
+                    }
                     continue;
                 }
